drm/rockchip: lvds: using readable connector name when dual lvds enabled
When we enable Dual LVDS, there will be the following
output in the VOP summary node:

$ cat /sys/kernel/debug/dri/0/summary
...
Video Port1: ACTIVE
    Connector: LVDS-1
...

The connector information here is not clear enough. When Dual LVDS is enabled,
both LVDS-0 and LVDS-1 will be enabled (but we will only expose one connector
to user space for operation, which can shield the underlying differences).

In addition, if only LVDS1 is enabled instead of using Dual LVDS,
"LVDS-1" will also be printed here.

In order to express this information more accurately, we will change
the corresponding connector name to "LVDS-DUAL" when using Dual LVDS:

$ cat /sys/kernel/debug/dri/0/summary
...
Video Port1: ACTIVE
    Connector: LVDS-DUAL
...

Change-Id: I9621338c8bdec4b32772c3558d55b5b426d3ccc4
Signed-off-by: Chaoyi Chen <chaoyi.chen@rock-chips.com>