thermal: cpu_cooling: resolve issue with CPU frequency limiting
The thermal_of_cooling_device_register function calls the
power_allocator_throttle function, which obtains a 'cur_state'. Based on this
'cur_state', it looks up the frequency table to assign a value to
cpufreq_cdev->clipped_freq. After registration is complete,
cpufreq_cdev->clipped_freq is initialized to the maximum value in the
freq_table, which overwrites the assignment in the power_allocator_throttle
function.

Referring to the solution in the 5.10 kernel code, move the code that
initializes cpufreq_cdev->clipped_freq to before the registration function.

Commit 5130802ddbb10a73cf17d1117e2044c8d5ba7d65 ("thermal: cpu_cooling: Switch
to QoS requests for freq limits") Switch over to using the QoS request for
maximum frequency constraint for cpu_cooling driver.

Signed-off-by: Ye Zhang <ye.zhang@rock-chips.com>
Change-Id: I48e8fa9237ac591a3d86f08242aa48ff73e14411